feat: :arrow_up: Update rust-sugiyama crate.

+++ opossum_gui/src/components/scenery_editor/constants.rs
@@ -19,7 +19,7 @@ pub const PORT_HEIGHT: f64 = 13.0;
 pub const PORT_WIDTH: f64 = 13.0;
 
 // constants for GraphStore
-pub const SUGIYAMA_VERTEX_SPACING: usize = 250;
+pub const SUGIYAMA_VERTEX_SPACING: f64 = 250.0;
 pub const SUGIYAMA_VERT_PATH_FACTOR: f64 = 0.7;
 
 // constant for node positioning
